requirements.txt


# FastAPI + server
fastapi==0.115.0
uvicorn[standard]==0.30.0

# Google Gemini API
google-generativeai==0.7.2
python-dotenv==1.0.1

# Excel handling
openpyxl==3.1.5

# Data & validation
pydantic==2.9.2

# CORS
starlette==0.37.2

# (Optional but useful)
httpx==0.27.0        # async requests if needed later
requests==2.32.3     # standard requests
